pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "fivb-leaderboard"
version = "0.1.0"
description = "Semantic data model for FIVB Beach Pro Tour leaderboards (dbt + Postgres)"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "dbt-postgres>=1.8.0,<2.0.0",
    "requests>=2.32.0,<3.0.0",
    "SQLAlchemy>=2.0.0,<3.0.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pytest>=7.0.0,<9.0.0",
    "tqdm>=4.66.0,<5.0.0",
    "streamlit>=1.28.0,<2.0.0",
    "plotly>=5.18.0,<6.0.0",
    "pandas>=2.0.0,<3.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["etl*"]
